qua-types: rename module and use submodules
